% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linbin.R
\name{as.event.table}
\alias{as.event.table}
\alias{as.event.table.data.frame}
\alias{as.event.table.matrix}
\alias{as.event.table.numeric}
\title{Coerce to an Event Table}
\usage{
as.event.table(x, ...)

\method{as.event.table}{numeric}(x)

\method{as.event.table}{matrix}(x, from.col = "from", to.col = "to")

\method{as.event.table}{data.frame}(x, from.col = "from", to.col = "to")
}
\arguments{
\item{x}{object to be coerced to an event table.}

\item{from.col,to.col}{names or indices of the columns in \code{x} containing the event endpoints. Values are swapped as needed to ensure that \code{to ≥ from} on all rows.}
}
\value{
An event table, the \code{data.frame} object used by \code{linbin} to describe interval data.
}
\description{
Attempts to coerce an object to an event table.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: Expands a numeric vector into two columns of event endpoints: \code{(x1, x2), (x2, x3), (x3, x4)}.

\item \code{matrix}: Converts the matrix to a data frame, then calls the \code{data.frame} method.

\item \code{data.frame}: Renames \code{from.col} and \code{to.col} to "from" and "to" as needed. Since these column names must be unique, other columns cannot also be called "from" or "to".
}}
\examples{
as.event.table(1)
as.event.table(1:10)
as.event.table(cbind(1:10, 1:10), from.col = 1, to.col = 2)
as.event.table(data.frame(from = 1:10, to = 1:10))
}
\seealso{
\code{\link{event.table}}, \code{\link{read.event.table}} for reading files as event tables.
}

