% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linbin.R
\name{crop.events}
\alias{crop.events}
\title{Crop Events to Bins}
\usage{
crop.events(events, bins, scaled.cols = NULL)
}
\arguments{
\item{events}{object of class \code{EventTable}.}

\item{bins}{object of class \code{EventTable} containing only non-overlapping line events.}

\item{totals}{names of the columns to be interpreted as totals. As these are proportional to the event length, they must be recalculated following event cutting.}

\item{verbose}{logical value indicating whether to print verbose messages.}
}
\value{
An \code{EventTable} with the same columns as \code{events}.
}
\description{
\code{crop.EventTable} crops events to the specified bins.
}
\details{
Bins are an \code{EventTable} containing only non-overlapping line events, a requirement validated by calling \code{\link{validate.binEventTable}}. Events are cut at bin endpoints by calling \code{\link{cut.EventTable}}, but no other resampling is performed. Partial or whole events lying outside of the bins are removed. FIXME: (Points lying on bin endpoints are preserved.) Not anymore!
}
\seealso{
\code{\link{EventTable}}, \code{\link{is.binEventTable}}, \code{\link{cut.EventTable}}.
}

