% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linbin.R
\name{cut.events}
\alias{cut.events}
\title{Cut Events}
\usage{
\method{cut}{events}(events, cuts, scaled.cols = NULL)
}
\arguments{
\item{events}{an event table.}

\item{cuts}{positions at which to cut the events. May be either a numeric vector or an event table.}

\item{scaled.cols}{names or indices of the columns to be rescaled after cutting.}
}
\value{
An \code{EventTable} with the same columns as \code{events}.
}
\description{
\code{cut.events} splits events apart at the specified \code{cuts} positions.
}
\details{
Line events straddling any of the \code{cutpoints} are duplicated as needed, and their \code{from} and \code{to} values upated to reflect the endpoints of the new line segments. Numeric columns named or numbered in \code{scaled.args} are scaled by the fraction of the original event length in each resulting segment (which assumes that these variables are uniformly distributed over the original interval). All other fields are left unchanged. To track the parents of the resulting event segments, add a unique ID field to \code{events} before calling \code{cut.EventTable}.
}
\seealso{
\code{\link{EventTable}}, \code{\link{crop.EventTable}}.
}

