% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\name{elwha}
\alias{elwha}
\title{Elwha River Survey}
\format{A data frame with 249 rows and 33 variables.}
\source{
Brenkman, S. J., J. J. Duda, C. E. Torgersen, E. Z. Welty, G. R. Pess, R. Peters, and M. L. McHenry. 2012. A riverscape perspective of Pacific salmonids and aquatic habitats prior to large-scale dam removal in the Elwha River, Washington, USA. Fisheries Management and Ecology 19:36-53. DOI: \href{http://dx.doi.org/10.1111/j.1365-2400.2011.00815.x}{10.1111/j.1365-2400.2011.00815.x}
}
\description{
An event table containing the results of a survey of the Elwha River (Washington, USA) carried out in August-September 2008. Both physical variables and fish counts were collected.
}
\details{
\itemize{
  \item from, to - distance upstream from the river mouth [km]
  \item unit.length - unit length [m]
  \item unit.type - unit type (P = pool, GP = glide-like pool, G = glide, GR = glide-like riffle, R = riffle)
  \item channel.type - channel type (1 = main, 2 = secondary)
  \item mean.depth - mean depth [m]
  \item max.depth - max depth [m]
  \item mean.width - mean wetted width [m]
  \item bedrock - bedrock substrate [\%]
  \item boulder - boulder substrate [\%]
  \item cobble - cobble substrate [\%]
  \item gravel - gravel substrate [\%]
  \item sand - sand substrate [\%]
  \item silt - silt substrate [\%]
  \item overhang.cover - channel banks with overhanging vegetation [\%]
  \item boulder.cover - channel area covered by boulders [\%]
  \item jams - number of log jams
  \item jam.area - total area of log jams [m^2]
  \item SACO.10/20/30/40/total - number of Bull Trout (Salvelinus confluentus) sized 10 - 20 cm / 20 - 30 cm / 30 - 40 cm / > 40 cm / total, respectively.
  \item ONXX.10/20/30/40/total - number of trout (Oncorhynchus sp.) sized 10 - 20 cm / 20 - 30 cm / 30 - 40 cm / > 40 cm / total, respectively.
  \item SAFO - number of Brook Trout (Salvelinus fontinalis)
  \item ONTS - number of Chinook Salmon (Oncorhynchus tshawytscha)
  \item ONNE - number of Sockeye Salmon (Oncorhynchus nerka)
  \item LATR - number of Pacific Lamprey (Lampetra tridentata)
  \item ONKI - number of Coho Salmon (Oncorhynchus kisutch)
}
}

