% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linbin.R
\name{event.coverage}
\alias{event.coverage}
\title{Event Coverage}
\usage{
event.coverage(events, closed = TRUE)
}
\arguments{
\item{events}{object of class \code{EventTable}.}

\item{closed}{logical value indicating whether events should be interpreted as closed intervals. If \code{TRUE} (the default), coverage is continuous at breaks between two adjacent but non-overlapping events.}

\item{verbose}{logical value indicating whether to print verbose messages.}
}
\value{
\code{coverage.EventTable} and \code{gaps.EventTable} both return an \code{EventTable} containing only event endpoints (\code{FROM} and \code{TO}). If no gaps are present, \code{gaps.EventTable} returns \code{NA}.
}
\description{
\code{coverage.EventTable} returns the flattened coverage, or union, of all the events in an \code{EventTable}. In other words, the result consists of the intervals, of maximal length, over which the number of events is always 1 or greater. Conversely, \code{gaps.EventTable} returns the gaps in event coverage; the intervals, of maximal length, over which there are 0 events. FIXME: Drops points on boundaries, even when closed = TRUE. Drops isolated points with closed = FALSE. Correct behavior?
}
\seealso{
\code{\link{EventTable}}, \code{\link{range.EventTable}} for event range (coverage with gaps ignored).
}

