% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linbin.R
\name{event.range}
\alias{event.range}
\title{Event Range}
\usage{
event.range(events)
}
\arguments{
\item{events}{an event table.}
}
\value{
A single-row event table.
}
\description{
\code{event.range} returns a single-row event table with the minimum and maximum endpoints of all the events in an event table.
}
\seealso{
\code{\link{event.coverage}}, \code{\link{event.gaps}}
}

