% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linbin.R
\name{event.table}
\alias{event.table}
\title{Event Tables}
\usage{
event.table(from = integer(), to = integer(), ...)
}
\arguments{
\item{from,to}{event endpoints, in any format coercible to single data frame columns. \code{from} and \code{to} are swapped as needed so that \code{to} ≥ \code{from} on all rows. If only \code{from} is passed, \code{\link{as.event.table}} is dispatched for object coercion.}

\item{...}{additional arguments, which can be either of the form \code{value} or \code{tag = value}, to be  passed directly to \code{\link{data.frame}} following \code{from} and \code{to}. Component names are created based on the tag (if present) or the deparsed argument itself.}
}
\value{
An event table, the \code{data.frame} object used by \code{linbin} to describe interval data.
}
\description{
This function creates an event table, a custom \code{data.frame} used throughout the \code{linbin} package to store and manipulate linearly referenced data. Each row includes an event's endpoints \code{from} and \code{to} (which can be equal, to describe a point, or non-equal, to describe a line) and the values of any variables measured on that interval.
}
\details{
Event endpoints (and any additional arguments) are coerced to a data frame by calling \code{\link{data.frame}}, then coerced to an event table by calling \code{\link{as.event.table}}. A valid event table has two columns named "from" and "to" containing only finite numeric values (i.e., no \code{NA}, \code{NaN}, or \code{Inf}) and ordered such that \code{to} ≥ \code{from}. \code{\link{is.event.table}} tests for these requirements. The other columns in the \code{event.table} can be of any type supported by the \code{data.frame} class. The core routines of \code{linbin} — particularly \code{\link{sample.events}} and \code{\link{plot.events}} — rely on "bins", event tables consisting of only non-overlapping, line events (\code{\link{is.bin.event.table}}) to summarize and plot event data.
}
\examples{
event.table(1, 10)
event.table(1:10)
event.table(c(0, 15, 25), c(10, 30, 35), x = 1, f = c('a', 'b', 'c'))
}
\seealso{
\code{\link{data.frame}}.

\code{\link{as.event.table}} and \code{\link{import.event.table}} for coercing objects and files to event tables, \code{\link{is.event.table}} to validate event tables, and \code{\link{seq.bin.event.table}} and \code{\link{is.bin.event.table}} to generate and validate bin event tables.
}

