% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_metrics.R
\name{event_midpoints}
\alias{event_midpoints}
\title{Event midpoints}
\usage{
event_midpoints(e)
}
\arguments{
\item{e}{Event table.}
}
\description{
Event midpoints
}
\examples{
e <- events(c(0, 10, 15, 25, 30), c(10, 20, 25, 40, 30))
event_midpoints(e)
}
