% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/event_metrics.R
\name{event_range}
\alias{event_range}
\title{Event Range}
\usage{
event_range(e)
}
\arguments{
\item{e}{An event table.}
}
\description{
Returns the minimum and maximum endpoints of all the events in an event table.
}
\examples{
event_range(events(1:5))            # no gaps
event_range(events(c(1,5), c(1,5))) # gaps
}
\seealso{
\code{\link{event_coverage}} for an alternative that accounts for gaps.
}

