% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linbin.R
\name{find.intersecting}
\alias{find.intersecting}
\title{Find Intersecting Events}
\usage{
find.intersecting(x, y, closed.from = FALSE, closed.to = FALSE)
}
\arguments{
\item{x,y}{event tables.}

\item{closed.from,closed.to}{logical values indicating whether to treat events in \code{x} as closed intervals at their from and to endpoints, respectively. If \code{FALSE} (the default), events sharing an endpoint is NOT reported as an intersection.}
}
\description{
\code{find.intersecting} returns a logical matrix of intersections for all combinations of events in \code{x} and \code{y}.
}
\seealso{
\code{\link{cut.events}}, \code{\link{sample.events}}.
}

