% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{if_else}
\alias{if_else}
\title{Generalized If Else}
\usage{
if_else(test, yes, no)
}
\arguments{
\item{test}{an object which can be coerced to a logical value.}

\item{yes}{value returned if \code{test} is \code{TRUE}}

\item{no}{value returned if \code{test} is \code{FALSE}}
}
\description{
Returns different values depending on whether a test is \code{TRUE} or \code{FALSE}.
}
\seealso{
\code{\link{ifelse}}
}
\keyword{internal}

