% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linbin.R
\name{is.event.table}
\alias{is.bin.event.table}
\alias{is.event.table}
\title{Tests for Event Table Types}
\usage{
is.event.table(x, verbose = FALSE)

is.bin.event.table(x, group = NULL, verbose = FALSE)
}
\arguments{
\item{x}{any R object.}

\item{verbose}{logical value indicating whether to print the reason for test failure.}

\item{group}{a vector defining the bin groups. If \code{NULL}, the bins are treated as one group.}
}
\value{
Returns \code{TRUE} if \code{x} meets the criteria of the test, and \code{FALSE} otherwise.
}
\description{
\code{is.event.table} tests whether the object meets the requirements of an event table. \code{is.bin.event.table} tests all these, plus whether the event table contains only groups of non-overlapping line events.
}
\seealso{
\code{\link{event.table}}, \code{\link{as.event.table}}.
}

