% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/event_tests.R
\name{is_bin_events}
\alias{is_bin_events}
\title{Validate Bin Event Table}
\usage{
is_bin_events(x, group.col = NULL, verbose = FALSE)
}
\arguments{
\item{x}{an R object.}

\item{group.col}{name or numeric index of the column defining the groupping of bins. If \code{NULL}, the bins are treated as one group.}

\item{verbose}{logical value indicating whether to print the reason for test failure.}
}
\description{
Tests whether an object is an event table, plus whether the event table contains only groups of non-overlapping line events. FIXME: Deprecated, since only one function still actually requires this (\code{\link{seq_bin_events}}).
}
\seealso{
\code{\link{seq_bin_events}} and \code{\link{group_nonoverlapping_events}} for creating valid bin event tables.
}
\keyword{internal}

