% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{is_not_integer}
\alias{is_not_integer}
\title{Not Integers}
\usage{
is_not_integer(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
\code{FALSE} if \code{x} is a numeric vector containing only whole numbers, \code{TRUE} otherwise.
}
\description{
Tests whether a vector is empty, non-numeric, or contains any non-integer numbers.
}
\keyword{internal}

