% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\name{netmap}
\alias{netmap}
\title{Dungeness River (NetMap)}
\format{A data frame with 16,616 rows and 47 variables.}
\source{
\url{http://www.terrainworks.com/netmap-demo-tools-download}
}
\description{
NetMap (\href{http://www.terrainworks.com/}{terrainworks.com}) output for the entire fluvial network of the Dungeness River (Washington, USA). NetMap employs digital elevation models to generate detailed river networks and compute biophysical variables for spatially continuous hydrologic units throughout the networks.
}
\details{
\itemize{
  \item CHAN_ID - channel identifier (1 = mainstem, all others are tributaries)
  \item OUT_DIST - distance upstream from the river mouth to the downstream end of the unit [km]
  \item LENGTH_M - unit length [m]
  \item ... - see NetMap's \href{http://www.netmaptools.org/Pages/NetMapHelp/master_attribute_list.htm}{Master Attribute List}
}
}

