% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{quinault}
\alias{quinault}
\title{Quinault River Survey}
\format{A data frame with 363 rows and 31 variables.}
\source{
Samuel J. Brenkman (National Park Service, Olympic National Park, Washington, USA), unpublished data.
}
\description{
An event table containing the results of a survey of the Quinault River (Washington, USA) in August 2009. Both physical variables and fish counts were collected.
}
\details{
\itemize{
  \item from, to - distance upstream from the river mouth [km]
  \item altitude - mean elevation above sea level [m]
  \item channel.type - channel type (1 = main, 2 = secondary)
  \item unit.type - unit type (P = pool, GP = glide-like pool, GR = glide-like riffle, R = riffle)
  \item unit.length - unit length [m]
  \item mean.width - mean wetted width [m]
  \item mean.depth - mean depth [m]
  \item max.depth - max depth [m]
  \item overhang.cover - channel banks with overhanging vegetation [\%]
  \item boulder.cover - channel area covered by boulders [\%]
  \item jams - number of log jams
  \item jam.area - total area of log jams [m^2]
  \item SACO.10/20/30/50/total - number of Bull Trout (\emph{Salvelinus confluentus}) sized 10 - 20 cm / 20 - 30 cm / 30 - 50 cm / > 50 cm / total, respectively.
  \item ONXX.10/20/30/total - number of trout (\emph{Oncorhynchus sp.}) sized 10 - 20 cm / 20 - 30 cm / > 30 cm / total, respectively.
  \item PRWI - number of Mountain Whitefish (\emph{Prosopium williamsoni})
  \item ONTS - number of Chinook Salmon (\emph{Oncorhynchus tshawytscha})
  \item ONMY - number of Rainbow Trout (\emph{Oncorhynchus mykiss})
  \item ONKI - number of Coho Salmon (\emph{Oncorhynchus kisutch})
  \item ONNE - number of Sockeye Salmon (\emph{Oncorhynchus nerka})
  \item ONGO - number of Pink Salmon (\emph{Oncorhynchus gorbushcha})
  \item ONKE - number of Chum Salmon (\emph{Oncorhynchus keta})
  \item CAMA - number of Largescale Sucker (\emph{Catostomus macrocheilus})
  \item LATR - number of Pacific Lamprey (\emph{Lampetra tridentata})
}
}

