% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rgrep_exact}
\alias{rgrep_exact}
\title{Full Name Matching}
\usage{
rgrep_exact(patterns, x, ..., simplify = FALSE)
}
\arguments{
\item{patterns}{Character vector containing regular expressions to be matched.}

\item{x}{Character vector where matches are sought, or an object which can be coerced by \code{as.character} to a character vector.}

\item{...}{Additional arguments to \code{\link{grep}}.}

\item{simplify}{If \code{TRUE}, the result is simplified from a list to a vector or matrix if appropriate.}
}
\value{
List, matrix, or vector of the indices of the elements of \code{x} that yielded a match to each element of \code{patterns}.
}
\description{
Recursive wrapper for \code{grep} returning only full matches to elements of a character vector.
}
\keyword{internal}

