% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sample_events.R
\name{sampling_functions}
\alias{sampling_functions}
\title{Build Sampling Functions}
\usage{
sampling_functions(col.names, ...)
}
\arguments{
\item{col.names}{Names of columns, for converting column name indices to numeric colum indices.}

\item{...}{Lists of sampling function parameters (see \code{\link{sample_events}}).}
}
\value{
A list of functions.
}
\description{
Helper function for \code{\link{sample_events}}. Parses function call parameters into self-enclosed function calls that can be passed row subsets of an event table for sampling.
}
\details{
NOTE: Assumes bin assignments will be appended to the end of the event table.
}
\keyword{internal}

