% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{simple}
\alias{simple}
\title{Simple Event Table}
\format{A data frame with 11 rows and 7 variables.}
\description{
A simple, hypothetical event table.
}
\details{
\itemize{
  \item from, to - endpoint positions
  \item x, y, z - numeric variables
  \item factor - a factor variable
}
}

