% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linbin.R
\name{sort.events}
\alias{is.not.sorted}
\alias{sort.events}
\title{Sort Events}
\usage{
sort.events(events)

is.not.sorted(events)
}
\arguments{
\item{events}{an event table.}
}
\description{
Sorts events by ascending \code{from}, then ascending \code{to}. The function \code{is.not.sorted} quickly checks whether the events are not already sorted, without the cost of sorting them.
}

