% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linbin.R
\name{transform.events}
\alias{transform.events}
\title{Transform Events}
\usage{
\method{transform}{events}(events, scale = 1, translate = 0)
}
\arguments{
\item{events}{an event table.}

\item{scale}{number by which event endpoints should be scaled.}

\item{translate}{number by which event endpoints should be translated.}
}
\value{
An event table with scaled and translated \code{from} and \code{to} columns.
}
\description{
\code{transform.events} transforms events by scaling, then translating their endpoint positions. That is, the transformed \code{[from, to] = scale * [from, to] + translate}.
}

