% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/event_operations.R
\name{transform_events}
\alias{transform_events}
\title{Transform Events}
\usage{
transform_events(e, scale = 1, translate = 0)
}
\arguments{
\item{e}{An event table.}

\item{scale}{Number by which event endpoints should be scaled.}

\item{translate}{Number by which event endpoints should be translated.}
}
\description{
Transforms events by scaling, then translating their endpoint positions. That is, the transformed \code{[from, to] = scale * [from, to] + translate}.
}
\examples{
e <- events(c(10, 100), c(100, 1000))
transform_events(e, scale = 2, translate = 1)
}

